pyproject.toml
[project]
name = "kursach"
version = "0.1.0"
description = ""
authors = [
    {name = "nevek12",email = "vladnevvk@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "aiohttp (>=3.11.14,<4.0.0)",
    "langchain-ollama (>=0.3.0,<0.4.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
